reorganize aziotctl system commands
