+++ aziotctl/aziotctl-common/src/check_last_modified.rs
@@ -3,23 +3,24 @@
 use std::fs;
 use std::path::Path;
 
-pub enum CheckResult {
-    Ok,
+use crate::program_name;
+
+pub enum LastModifiedError {
     Ignored,
     Warning(String),
     Failed(std::io::Error),
 }
 
-pub fn check_last_modified(services: &[&str]) -> CheckResult {
+pub fn check_last_modified(services: &[&str]) -> Result<(), LastModifiedError> {
     let config = Path::new("/etc/aziot/config.toml");
 
     if !config.exists() {
-        return CheckResult::Ignored;
+        return Err(LastModifiedError::Ignored);
     }
 
     let config_metadata = match fs::metadata(config) {
         Ok(m) => m,
-        Err(err) => return CheckResult::Failed(err),
+        Err(err) => return Err(LastModifiedError::Failed(err)),
     };
 
     let config_last_modified = config_metadata
@@ -31,17 +32,17 @@ pub fn check_last_modified(services: &[&str]) -> CheckResult {
         let service_config = Path::new(&service_config);
 
         if !service_config.exists() {
-            return CheckResult::Warning(format!(
+            return Err(LastModifiedError::Warning(format!(
                 "{} does not exist.\n\
                 Did you run '{} config apply'?",
                 service_config.display(),
-                super::program_name()
-            ));
+                program_name()
+            )));
         }
 
         let service_config_metadata = match fs::metadata(service_config) {
             Ok(m) => m,
-            Err(err) => return CheckResult::Failed(err),
+            Err(err) => return Err(LastModifiedError::Failed(err)),
         };
 
         let service_config_last_modified = service_config_metadata
@@ -49,16 +50,16 @@ pub fn check_last_modified(services: &[&str]) -> CheckResult {
             .expect("file metadata should contain valid last_modified");
 
         if config_last_modified > service_config_last_modified {
-            return CheckResult::Warning(format!(
+            return Err(LastModifiedError::Warning(format!(
                 "{} was modified after {}'s config\n\
                 You must run '{} config apply' to update {}'s config with the latest config.toml",
                 config.display(),
                 service,
-                super::program_name(),
+                program_name(),
                 service
-            ));
+            )));
         }
     }
 
-    CheckResult::Ok
+    Ok(())
 }
