+++ aziotctl/aziotctl-common/src/system/system_logs.rs
