+++ aziotctl/src/internal/check/checks/up_to_date_configs.rs
@@ -3,8 +3,7 @@
 use anyhow::anyhow;
 use serde::Serialize;
 
-use aziotctl_common::check_last_modified::check_last_modified;
-use aziotctl_common::check_last_modified::CheckResult as InternalCheckResult;
+use aziotctl_common::check_last_modified::{check_last_modified, LastModifiedError};
 
 use crate::internal::check::{CheckResult, Checker, CheckerCache, CheckerMeta, CheckerShared};
 
@@ -22,10 +21,10 @@ impl Checker for UpToDateConfigs {
 
     async fn execute(&mut self, _shared: &CheckerShared, _cache: &mut CheckerCache) -> CheckResult {
         match check_last_modified(&["keyd", "certd", "identityd", "tpmd"]) {
-            InternalCheckResult::Ok => CheckResult::Ok,
-            InternalCheckResult::Ignored => CheckResult::Ignored,
-            InternalCheckResult::Warning(message) => CheckResult::Warning(anyhow!(message)),
-            InternalCheckResult::Failed(error) => CheckResult::Failed(error.into()),
+            Ok(()) => CheckResult::Ok,
+            Err(LastModifiedError::Ignored) => CheckResult::Ignored,
+            Err(LastModifiedError::Warning(message)) => CheckResult::Warning(anyhow!(message)),
+            Err(LastModifiedError::Failed(error)) => CheckResult::Failed(error.into()),
         }
     }
 }
